Adding parent and toString();
Creating a way to represent the elements with a string to see the structure of the program.

Also adding a way to get the parent of a card node because I couldn't figure out a way around it. Specifically for click and drag for solitare-like movement; being able to reference all the other children of the base tuple